Fix visit_definition edge case in DimensionRecord initialization.
Before we added visit_definition, it was safe to assume that taking
a DimensionElement's required dimensions, expanding them, and then
removing any dimension implied by another would yield the element's
required dimensions.  That's no longer always true, and that's
revealed a lot of fragile code that depended on the ordering of
dependent dimensions in dimension records.