Make some fixes for mypy
The fundamental problem is that the new row for the
composite formatter has no file associated with it.
This means that the Location is None so we have to
do lots of checking for mypy for that condition.
It also means that we can not construct a Formatter
(since it requires a Location).

This leads to recording the formatter class as well
as the formatter in the dataclass we use for collecting
all the relevant information. The formatter can now be
optional.

This leads to one of the Formatter methods becoming
a class method.